==> 2-run vida code to get excels of data.R <==

createAllFilesFromAllVideos(inPath ="//medixsrv/Nematodes/data",outPath = "C://Users/lhuang37/Downloads/40 test 20160115", 
                            wormList = c("n2_f","n2_nf","n2_nnf","tph1_f","tph1_nf","egl19_nf","che2_f","che2_nf",
                                         "egl19_f","egl19h20_nf"), delta = 40)
